+++ CHANGELOG.md
@@ -4,7 +4,21 @@ All notable changes to this repo will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/), 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.2] - In 2019-06-25
+## [1.0.3] - 2020-03-27
+### Added
+- Helpers for controlling logging options when a command line is not
+    available, such as in notebooks or from test cases
+- Helpers for pytorch, including early stopping based on validation set
+    performance
+
+### Updated
+- Binary ML metrics to include precision-recall curves
+- Optional progress bar when collecting dask results
+
+### Fixed
+- Safely load YAML config files
+
+## [1.0.2] - 2019-06-25
 ### Added
 - Additional documentation
 - Helper for plotting binary prediction scores